[master] Removed redundant code.
I wrote this to add padding to the plot. However, things were not that simple manipulating the plot padding in MATLAB. After some tests, this proved to be redundant. The other reason I removed this is to remove the following confusion: "Does the assignments of these values affect the result of 'plot_linear'? Are they necessary?".